refactor: ディレクトリ構成を変更 :building_construction:
