+++ package/model/lib/src/feature/admob/_exporter.dart
@@ -1,0 +1,3 @@
+export 'interstitial_ad_service.dart';
+export 'rewarded_ad_notifier.dart';
+export 'rewarded_ad_state.dart';
