--- package/model/lib/src/feature/admob/admob.dart
@@ -1,3 +1,0 @@
-export 'application/interstitial_ad_service.dart';
-export 'application/rewarded_ad_notifier.dart';
-export 'domain/rewarded_ad_state.dart';
