+++ package/model/lib/src/feature/admob/consecutive_ad_count_notifier.dart
