+++ package/model/lib/src/feature/admob/rewarded_ad_notifier.dart
@@ -2,7 +2,7 @@ import 'package:common/common.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../domain/rewarded_ad_state.dart';
+import 'rewarded_ad_state.dart';
 
 part 'rewarded_ad_notifier.g.dart';
