+++ package/model/lib/src/feature/admob/rewarded_ad_state.freezed.dart
